// web/src/gen/model/pinPrimitiveType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type PinPrimitiveType =
  (typeof PinPrimitiveType)[keyof typeof PinPrimitiveType];

export const PinPrimitiveType = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
} as const;