// web/src/gen/model/pinCardinality.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */

export type PinCardinality =
  (typeof PinCardinality)[keyof typeof PinCardinality];

export const PinCardinality = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
} as const;